+++ src/ai-bundle/config/options.php
@@ -44,6 +44,7 @@ return static function (DefinitionConfigurator $configurator): void {
                     ->append($import('platform/gemini'))
                     ->append($import('platform/generic'))
                     ->append($import('platform/huggingface'))
+                    ->append($import('platform/inworld'))
                     ->append($import('platform/lmstudio'))
                     ->append($import('platform/mistral'))
                     ->append($import('platform/ollama'))